00_RESTful URL setting

+++ boards/views.py
@@ -10,14 +10,15 @@ def index(request):
     return render(request, 'boards/index.html', context)
     
 def new(request):
-    return render(request, 'boards/new.html')
-    
-def create(request):
-    title = request.POST.get('title')
-    content = request.POST.get('content')
-    board = Board(title=title, content=content)
-    board.save()
-    return redirect('boards:detail', board.pk)
+    # create 방식
+    if request.method == "POST":
+        title = request.POST.get('title')
+        content = request.POST.get('content')
+        board = Board(title=title, content=content)
+        board.save()
+        return redirect('boards:detail', board.pk)
+    else:
+        return render(request, 'boards/new.html')
     
 def detail(request, pk):
     board = Board.objects.get(pk=pk)
@@ -28,19 +29,21 @@ def detail(request, pk):
     
 def delete(request, pk):
     board = Board.objects.get(pk=pk)
-    board.delete()
-    return redirect('boards:index')
+    if request.method == 'POST':
+        board.delete()
+        return redirect('boards:index')
+    else:
+        return redirect('boards:detail', board.pk)
     
 def edit(request, pk):
     board = Board.objects.get(pk=pk)
-    context = {
-        'board' : board
-    }
-    return render(request, 'boards/edit.html', context)
-    
-def update(request, pk):
-    board = Board.objects.get(pk=pk)
-    board.title = request.POST.get('title')
-    board.content = request.POST.get('content')
-    board.save()
-    return redirect('boards:detail', board.pk)
+    if request.method == "POST":
+        board.title = request.POST.get('title')
+        board.content = request.POST.get('content')
+        board.save()
+        return redirect('boards:detail', board.pk)
+    else:
+        context = {
+            'board' : board
+        }
+        return render(request, 'boards/edit.html', context)
